# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Fehlerantwort (``HTTPException``): ``detail`` ist die Meldung.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SaveScenarioRequest": {
        "properties": {
          "config_yaml": {
            "title": "Config Yaml",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "config_yaml"
        ],
        "title": "SaveScenarioRequest",
        "type": "object"
      },
      "SavedScenario": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "default": "user",
            "enum": [
              "user",
              "example"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "SavedScenario",
        "type": "object"
      },
      "ScenarioContent": {
        "properties": {
          "config_yaml": {
            "title": "Config Yaml",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "config_yaml"
        ],
        "title": "ScenarioContent",
        "type": "object"
      },
      "ScenarioDeleted": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "deleted"
        ],
        "title": "ScenarioDeleted",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Token aus POST /api/auth/login als 'Authorization: Bearer <token>'. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft.",
        "scheme": "bearer",
        "type": "http"
      },
      "QueryToken": {
        "description": "Dasselbe Token als Query-Parameter - nur für Ereignisstrom, Rasterbild und Download, die der Browser ohne Authorization-Header abruft. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft.",
        "in": "query",
        "name": "token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend-API zur interaktiven Demonstration des GeoFACT-Frameworks.",
    "title": "GeoFACT Web API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scenarios": {
      "get": {
        "description": "Gespeicherte Szenarien (source=user, gemeinsamer Pool aller Nutzer) und die mitgelieferten Beispiele (source=example), je Gruppe nach Namen sortiert.",
        "operationId": "list_scenarios_api_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedScenario"
                  },
                  "title": "Response List Scenarios Api Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Szenarien auflisten",
        "tags": [
          "scenarios"
        ]
      },
      "post": {
        "description": "Speichert den YAML-Text unter dem Namen; ein vorhandenes Szenario gleichen Namens wird überschrieben. Der Text wird nicht geprüft.",
        "operationId": "save_scenario_api_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedScenario"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Der Name enthält kein verwertbares Zeichen."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Szenario speichern",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/scenarios/examples/{example_id}": {
      "get": {
        "description": "example_id ist der Pfad unter examples/ ohne Endung. Dem Text ist die Kopfzeile '# geofact-origin: <id>' vorangestellt, gegen die relative Pfade aufgelöst werden.",
        "operationId": "load_example_scenario_api_scenarios_examples__example_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "example_id",
            "required": true,
            "schema": {
              "title": "Example Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioContent"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Kein Beispiel mit dieser id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mitgeliefertes Beispiel laden",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/scenarios/{name}": {
      "delete": {
        "description": "Ein unbekannter Name ist kein Fehler: deleted ist dann false.",
        "operationId": "delete_scenario_api_scenarios__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDeleted"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Der Name enthält kein verwertbares Zeichen."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Gespeichertes Szenario löschen",
        "tags": [
          "scenarios"
        ]
      },
      "get": {
        "description": "Liefert den gespeicherten YAML-Text.",
        "operationId": "load_scenario_api_scenarios__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioContent"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Der Name enthält kein verwertbares Zeichen."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Kein Szenario dieses Namens."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Gespeichertes Szenario laden",
        "tags": [
          "scenarios"
        ]
      }
    },
...
  },
...
}